+++ backend/src/main/java/com/example/backend/services/ProductService.java
@@ -43,7 +43,7 @@ public class ProductService {
         return productRepository.save(product, authIdClient);
     }
 
-    public ProductEntity updateProduct(Long id, ProductEntity product) {
+    public ProductEntity updateProduct(int id, ProductEntity product) {
         ProductEntity possibleProduct = productRepository.findById(id);
         int authIdClient = authService.getAuthIdClient();
         if (possibleProduct == null) {
